requirements.txt
# Dotfiles setup requirements
PyYAML>=6.0